need long long int for double on 32 bits

+++ regression/cbmc/Pointer_byte_extract6/main.c
@@ -16,7 +16,7 @@ int main()
   assert(*(unsigned int *)&signed_int==1);
   assert(((struct some *)&signed_int)->f==1);
   assert(*(int *)&a_float==1065353216);
-  assert(*(long int *)&a_double==4607182418800017408l);
+  assert(*(long long int *)&a_double==4607182418800017408l);
   
   // other direction
   signed_int=1065353216;
